LPD-53900 Moves class to the correct package
