+++ modules/apps/style-book/style-book-item-selector-api/bnd.bnd
@@ -1,4 +1,4 @@
 Bundle-Name: Liferay Style Book Item Selector API
 Bundle-SymbolicName: com.liferay.style.book.item.selector.api
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.style.book.item.selector.criterion
+Export-Package: com.liferay.style.book.item.selector
