+++ modules/apps/style-book/style-book-item-selector-api/src/main/resources/com/liferay/style/book/item/selector/packageinfo
